+++ contracts/utils/ITokenUtils.sol
@@ -3,9 +3,9 @@ pragma solidity ^0.8.19;
 
 // Author: Francesco Sullo <francesco@sullo.co>
 
-//import "hardhat/console.sol";
-import "../vaults/IFlexiVault.sol";
-import "../protected-nft/IActors.sol";
+//import {console} from "hardhat/console.sol";
+import {IFlexiVault} from "../vaults/IFlexiVault.sol";
+import {IActors} from "../protected-nft/IActors.sol";
 
 interface ITokenUtils {
   function isTokenUtils() external pure returns (bytes4);
@@ -23,7 +23,7 @@ interface ITokenUtils {
     uint256 amount,
     address beneficiary,
     uint256 timestamp,
-    uint validFor
+    uint256 validFor
   ) external view returns (bytes32);
 
   function hashWithdrawsRequest(
@@ -34,17 +34,17 @@ interface ITokenUtils {
     uint256[] memory amounts,
     address[] memory beneficiaries,
     uint256 timestamp,
-    uint validFor
+    uint256 validFor
   ) external view returns (bytes32);
 
-  function hashEjectRequest(uint256 owningTokenId, uint256 timestamp, uint validFor) external view returns (bytes32);
+  function hashEjectRequest(uint256 owningTokenId, uint256 timestamp, uint256 validFor) external view returns (bytes32);
 
   function hashRecipientRequest(
     address owner,
     address recipient,
     IActors.Level level,
     uint256 timestamp,
-    uint validFor
+    uint256 validFor
   ) external view returns (bytes32);
 
   function hashBeneficiaryRequest(
@@ -52,7 +52,7 @@ interface ITokenUtils {
     address beneficiary,
     IActors.Status status,
     uint256 timestamp,
-    uint validFor
+    uint256 validFor
   ) external view returns (bytes32);
 
   /**
@@ -62,5 +62,10 @@ interface ITokenUtils {
    * @param timestamp The timestamp of the transfer request
    * @return The hash of the transfer request
    */
-  function hashTransferRequest(uint256 tokenId, address to, uint256 timestamp, uint validFor) external view returns (bytes32);
+  function hashTransferRequest(
+    uint256 tokenId,
+    address to,
+    uint256 timestamp,
+    uint256 validFor
+  ) external view returns (bytes32);
 }
